[wip] move library outside toolbar
