+++ src/components/LockIcon.tsx
@@ -52,6 +52,7 @@ export const LockIcon = (props: LockIconProps) => {
         },
       )}
       title={`${props.title} — Q`}
+      style={{ marginRight: "var(--space-factor)" }}
     >
       <input
         className="ToolIcon_type_checkbox"
